+++ src/mkdir.ts
@@ -1,8 +1,7 @@
 import fs from 'fs';
 
 /** Recursively creates directories. */
-export const makeDirectories = async (path: string) => {
-  if (!fs.existsSync(path)) {
-    await fs.promises.mkdir(path, { recursive: true });
-  }
-};
+export const makeDirectories = async (path: string) =>
+  void (
+    fs.existsSync(path) || (await fs.promises.mkdir(path, { recursive: true }))
+  );