+++ src/remove_first_directory.ts
@@ -1,5 +1,0 @@
-import path from 'path';
-
-/** Discards first directory from _url_. */
-export const removeFirstDirectory = (url: string) =>
-  url.split(path.sep).slice(1).join(path.sep);